requirements.txt
-e git://github.com/fusionbox/django-fusionbox.git#egg=django_fusionbox
Django==1.4
South
Werkzeug
django-cachebuster
django-compressor
django-debug-toolbar
django-extensions
MySQL-python==1.2.3
Mezzanine==1.1.4
pyScss==1.1.3
django-treebeard==1.61
psycopg2==2.4.5
